fix stage preview
Stage preview has a few bugs when manually entering size values into the inputbox. When the slider-value is changed, the inputbox is updated, but not vice versa. I suspect the error lies here, since when entering a value manually, the slider is not updated and thus we have different values to work with. Giving us some funny outcomes.

Also there seems to be no easy solution to prevent the user from entering garbage without implementing user checks with regex (pattern=""). Patterns can only be used to check input, not to prevent input. Meaning the update process would need a refactor.

A quick and dirty workaround is to make the inputboxes read-only and increase the slider-step value so we can actually hit even values. Also the text-colors are changed to indicate read-only.